Initial restructure of configs
