+++ benchmarks/benchmark_base.py
@@ -18,7 +18,10 @@ from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
 from tardis.simulation import Simulation
 from tardis.tests.fixtures.atom_data import DEFAULT_ATOM_DATA_UUID
 from tardis.tests.fixtures.regression_data import RegressionData
-from tardis.transport.montecarlo import RPacket, montecarlo_configuration
+from tardis.transport.montecarlo import RPacket
+from tardis.transport.montecarlo.configuration.base import (
+    MonteCarloConfiguration,
+)
 from tardis.transport.montecarlo.estimators import radfield_mc_estimators
 from tardis.transport.montecarlo.numba_interface import opacity_state_initialize
 from tardis.transport.montecarlo.packet_collections import (
@@ -235,9 +238,7 @@ class BenchmarkBase:
 
     @property
     def verysimple_packet_collection(self):
-        return (
-            self.nb_simulation_verysimple.transport.transport_state.packet_collection
-        )
+        return self.nb_simulation_verysimple.transport.transport_state.packet_collection
 
     @property
     def nb_simulation_verysimple(self):
@@ -268,27 +269,19 @@ class BenchmarkBase:
 
     @property
     def verysimple_disable_line_scattering(self):
-        return (
-            self.nb_simulation_verysimple.transport.montecarlo_configuration.DISABLE_LINE_SCATTERING
-        )
+        return self.nb_simulation_verysimple.transport.montecarlo_configuration.DISABLE_LINE_SCATTERING
 
     @property
     def verysimple_continuum_processes_enabled(self):
-        return (
-            self.nb_simulation_verysimple.transport.montecarlo_configuration.CONTINUUM_PROCESSES_ENABLED
-        )
+        return montecarlo_globals.CONTINUUM_PROCESSES_ENABLED
 
     @property
     def verysimple_tau_russian(self):
-        return (
-            self.nb_simulation_verysimple.transport.montecarlo_configuration.VPACKET_TAU_RUSSIAN
-        )
+        return self.nb_simulation_verysimple.transport.montecarlo_configuration.VPACKET_TAU_RUSSIAN
 
     @property
     def verysimple_survival_probability(self):
-        return (
-            self.nb_simulation_verysimple.transport.montecarlo_configuration.SURVIVAL_PROBABILITY
-        )
+        return self.nb_simulation_verysimple.transport.montecarlo_configuration.SURVIVAL_PROBABILITY
 
     @property
     def static_packet(self):
@@ -359,10 +352,10 @@ class BenchmarkBase:
 
     @property
     def montecarlo_configuration(self):
-        return montecarlo_configuration.MonteCarloConfiguration()
+        return MonteCarloConfiguration()
 
     @property
-    def rpacket_tracker(self): 
+    def rpacket_tracker(self):
         return RPacketTracker(0)
 
     @property
@@ -395,7 +388,6 @@ class BenchmarkBase:
             v_inner=np.array([-1, -1], dtype=np.float64),
             v_outer=np.array([-1, -1], dtype=np.float64),
         )
-
 
     @property
     def estimators(self):
