+++ tardis/transport/montecarlo/configuration/base.py
@@ -20,8 +20,6 @@ numba_config_spec = [
     ("VPACKET_TAU_RUSSIAN", float64),
     ("INITIAL_TRACKING_ARRAY_LENGTH", int64),
     ("LEGACY_MODE_ENABLED", boolean),
-    ("ENABLE_RPACKET_TRACKING", boolean),
-    ("CONTINUUM_PROCESSES_ENABLED", boolean),
     ("VPACKET_SPAWN_START_FREQUENCY", float64),
     ("VPACKET_SPAWN_END_FREQUENCY", float64),
     ("ENABLE_VPACKET_TRACKING", boolean),
@@ -44,9 +42,6 @@ class MonteCarloConfiguration(object):
 
         self.INITIAL_TRACKING_ARRAY_LENGTH = 0
         self.LEGACY_MODE_ENABLED = False
-
-        self.ENABLE_RPACKET_TRACKING = False
-        self.CONTINUUM_PROCESSES_ENABLED = False
 
         self.VPACKET_SPAWN_START_FREQUENCY = 0
         self.VPACKET_SPAWN_END_FREQUENCY = 1e200
