+++ tardis/transport/montecarlo/configuration/montecarlo_globals.py
@@ -1,0 +1,2 @@
+ENABLE_RPACKET_TRACKING = False
+CONTINUUM_PROCESSES_ENABLED = False
